Relocating all the proxies to a subpackage
