+++ setup.py
@@ -34,7 +34,7 @@ setup(name='wpipe',
           "Topic :: Software Development :: Version Control :: Git"
       ],
       python_requires='>=3',
-      packages=['wpipe', 'wpipe.sqlintf', 'wpipe.scheduler'],
+      packages=['wpipe', 'wpipe.sqlintf', 'wpipe.proxies', 'wpipe.scheduler'],
       package_dir={'': 'src'},
       install_requires=['numpy', 'pandas', 'tenacity', 'tables', 'sqlalchemy',
                         'mysql-connector-python', 'mysqlclient', 'astropy', 'jinja2'],
