+++ src/test/data/tasks/make_targets.py
@@ -15,15 +15,21 @@ if __name__ == '__main__':
     for my_input in my_pipe.inputs:
         for my_dp in my_input.rawdataproducts:
             my_target = my_input.target(name=my_dp.filename)
+            my_target.options['Interactions'] = 0
             for my_conf in my_target.configurations:
-                my_target.options = {"Configuration '"+my_conf.name+"' completed": False}
+                my_target.options["Configuration '" + my_conf.name + "' completed"] = False
+                my_target.options['Interactions'] += 1
                 my_job.logprint("Starting target "+my_target.name+" config "+my_conf.name)
                 if 'WPIPE_NO_PBS_SCHEDULER' in os.environ.keys():
                     my_job.child_event('add_prefix', tag="config_id#"+str(my_conf.config_id),
                                        options={'config_id': my_conf.config_id}).fire()
                 else:
                     my_job.child_event('add_prefix', tag="config_id#"+str(my_conf.config_id),
-                                       options={'config_id': my_conf.config_id, 'submission_type': 'pbs', 'job_time': 5}).fire()
+                                       options={'config_id': my_conf.config_id,
+                                                'submission_type': 'pbs',
+                                                'job_time': 5}).fire()
                     my_job.child_event('add_prefix', tag="config_id2_#" + str(my_conf.config_id),
-                                       options={'config_id': my_conf.config_id, 'submission_type': 'pbs', 'job_time': 5}).fire()
+                                       options={'config_id': my_conf.config_id,
+                                                'submission_type': 'pbs',
+                                                'job_time': 5}).fire()
