+++ src/wpipe/Node.py
@@ -7,9 +7,9 @@ available in the main ``wpipe`` namespace - use that instead.
 """
 import socket
 from .core import datetime, si
-from .core import ChildrenProxy
 from .core import initialize_args, wpipe_to_sqlintf_connection, in_session
 from .core import split_path
+from .proxies import ChildrenProxy
 
 __all__ = ['Node']
