+++ src/wpipe/proxies/core.py
@@ -1,0 +1,40 @@
+#!/usr/bin/env python
+"""
+Contains the core import statements and developing tools of proxies
+
+Please note that this module is private. All functions and objects
+are available in the main ``wpipe.proxies`` namespace - use that instead.
+"""
+import ast
+import contextlib
+import sys
+import itertools
+import numbers
+import datetime
+
+import numpy as np
+
+from ..core import in_session
+
+__all__ = ['contextlib', 'sys', 'itertools', 'numbers', 'datetime', 'np',
+           'in_session', 'try_scalar']
+
+
+def try_scalar(string):
+    """
+    Returns given string parameter as scalar value, or return as string.
+
+    Parameters
+    ----------
+    string : string
+        Input string parameter.
+
+    Returns
+    -------
+    value
+        Scalar conversion from string or string itself.
+    """
+    try:
+        return ast.literal_eval(string)
+    except (ValueError, NameError, SyntaxError):
+        return string
